requirements.txt
anthropic
feedparser
requests
newspaper3k
lxml_html_clean
